universal_ingestion_agent/pyproject.toml
[project]
name = "ingestion-agent"
version = "0.1.0"
description = "Universal Ingestion Agent for multi-source data pipelines"
authors = [
    {name = "aakki"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.13"

[tool.poetry.dependencies]
python = "^3.13"
pandas = "^2.0.0"
minio = "^7.0.0"
requests = "^2.28.0"
pyyaml = "^6.0"
pydantic = "^1.10.0"
jinja2 = "^3.1.0"
requests-oauthlib = "^2.0.0"
fastparquet = "^2024.11.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-mock = "^3.10"
pytest-cov = "^4.0"
mypy = "^1.0"
black = "^23.0"
isort = "^5.12"

[project.scripts]
ingestion-agent = "ingestion_agent.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
